This is a rework of my original idea because ...
I thought I was doing it right, but I was going about it in a wrong way. This will allow the search of one city, and tell the user about the sighting. I am working on the shape of the ship and other categories. Use this one not the other .py